refactor: move iterator into new module (#112)
* refactor: move iterator into new module

* rename module

* fast but kinda broken

* not optimized but tests passing

* Revert "not optimized but tests passing"

This reverts commit 4bc76295a0d756b613963a9e433a05405b8b83a8.

* Revert "fast but kinda broken"

This reverts commit 6fb14f4dea9b773b083d939d35ac65e422b467c0.
