+++ src/useq/_utils.py
@@ -1,9 +1,35 @@
-from typing import Iterable, Iterator, Mapping, Tuple, TypeVar, Union
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Iterable, Iterator, Mapping, Tuple, TypeVar, Union
 
 from pydantic import validator
 
+if TYPE_CHECKING:
+    from typing import Final
+
 KT = TypeVar("KT")
 VT = TypeVar("VT")
+
+
+# could be an enum, but this more easily allows Axis.Z to be a string
+class Axis:
+    """Recognized axis names."""
+
+    TIME: Final[str] = "t"
+    POSITION: Final[str] = "p"
+    GRID: Final[str] = "g"
+    CHANNEL: Final[str] = "c"
+    Z: Final[str] = "z"
+
+
+# note: order affects the default axis_order in MDASequence
+AXES: Final[tuple[str, ...]] = (
+    Axis.TIME,
+    Axis.POSITION,
+    Axis.GRID,
+    Axis.CHANNEL,
+    Axis.Z,
+)
 
 
 class ReadOnlyDict(Mapping[KT, VT]):
